Temporary comment some uncompilable code pieces

+++ src/systems/rendering/engine/impl/opengl/OpenGlUtils.cpp
@@ -1,5 +1,7 @@
 #include "OpenGlUtils.h"
 
+#include <vector>
+
 #include <glbinding/gl/gl.h>
 
 namespace systems {
@@ -9,17 +11,17 @@ namespace opengl {
 namespace utils {
 
 ///// --------------------------------------------------------------- getInfoLog
-const std::string& getInfoLog(uint id) {
+const std::string& getInfoLog(unsigned int id) {
     gl::GLint maxLength = 0;
     gl::glGetShaderiv(id, gl::GL_INFO_LOG_LENGTH, &maxLength);
 
-    std::vector<char> errorLog(static_cast<uint64_t>(maxLength));
+    std::vector<char> errorLog(static_cast<unsigned int>(maxLength));
     gl::glGetShaderInfoLog(id, maxLength, &maxLength, &errorLog[0]);
 
     return std::string(errorLog.begin(), errorLog.end());
 }
 
-} // namespace utils
+}  // namespace utils
 }  // namespace opengl
 }  // namespace engine
 }  // namespace rendering
